Clip a multipoligon vector layer based on a polygon

We will use this function only if function `clip_vector()` gives an error. The error I have found is `ParseException: Unknown WKB type 12`, it has to do with type *MULTIPOLYGON*.

Usage

clip_multipoligon(vector, polygon)

Arguments

vector

A `sf` multipoligon vector layer.

polygon

A `sf` polygon layer.

Value

A `sf` vector layer.

Details

The solution adopted is based on the one described in https://gis.stackexchange.com/questions/389814/r-st-centroid-geos-error-unknown-wkb-type-12

We must ensure that the layer to be cut has the correct encoding.

Project the result to the CRS of the clipping polygon.

See also

Examples

#