Copy Layer Styles from Source to Destination in GeoPackage

Copies the first style definition from a source (either a GeoPackage file or a PostGIS connection) and assigns it to all layers in the destination GeoPackage.

Usage

copy_styles_layer(from, to)

Arguments

from

A data origin. This can be: - A string representing the path to a GeoPackage file. - A `DBI` database connection object to a PostGIS database, created using [RPostgres::dbConnect()].

to

A string representing the path to the destination GeoPackage file.

Value

The updated `layer_styles` table, returned invisibly.

Details

The function reads the first style from the `layer_styles` table in the source GeoPackage or PostGIS database. This style is then applied to all layers in the destination GeoPackage.

See also

Examples

if (FALSE) { # \dontrun{
source_gpkg <- "source.gpkg"
dest_gpkg <- "destination.gpkg"

copy_styles_layer(from = source_gpkg, to = dest_gpkg)
} # }