Copy Layer Styles from Source to Specific Layers in PostGIS Database

Copies the first style definition from a source (either a GeoPackage file or a PostGIS connection) to the specified layers in a PostGIS database.

Usage

copy_styles_layer_names(from, to, layers, database, schema = "public")

Arguments

from

A data origin. This can be: - A string representing the path to a GeoPackage file. - A `DBI` database connection object to a PostGIS database, created using [RPostgres::dbConnect()].

to

A database connection object to the destination PostGIS database (e.g., from `RPostgres::dbConnect`).

layers

A character vector of layer names in the destination database to which the style should be applied.

database

A string specifying the name of the PostGIS database.

schema

A string specifying the schema in the PostGIS database where the layers reside. Default is `"public"`.

Value

The updated `layer_styles` table, returned invisibly.

Details

The function reads the first style from the `layer_styles` table in the source and applies it to the specified layers in the destination PostGIS database.

See also

Examples

if (FALSE) { # \dontrun{
source_gpkg <- "source.gpkg"
conn <- DBI::dbConnect(
  RPostgres::Postgres(),
  dbname = "mydb",
  host = "localhost",
  user = "user",
  password = "password"
)

layers_to_style <- c("layer1", "layer2")

copy_styles_layer_names(
  from = source_gpkg,
  to = conn,
  layers = layers_to_style,
  database = "mydb",
  schema = "public"
)

DBI::dbDisconnect(conn)
} # }