GeoServer Connection Object (`geoserver` S3 Class)

This S3 class represents a connection to a GeoServer instance. It stores the connection details, including the base URL, user credentials, and the default workspace.

Usage

geoserver(url, user, password, workspace)

Arguments

url

A character string specifying the base URL of the GeoServer instance (e.g., `"http://localhost:8080/geoserver"`).

user

A character string representing the GeoServer username with the required permissions.

password

A character string representing the password for the specified user.

workspace

A character string specifying the default workspace to use in GeoServer operations.

Value

An object of class `geoserver` or NULL if an error occurred.

See also

Examples

if (FALSE) { # \dontrun{
gso <- geoserver(
  url = "http://localhost:8080/geoserver",
  user = "admin",
  password = "geoserver",
  workspace = "sigugr_test"
)
} # }