Publish a layer to GeoServer

Publishes a vector layer to GeoServer using the REST API. Returns a status code and prints an appropriate message indicating success or failure.

Usage

gs_publish_layer(
  geoserver_url,
  user,
  password,
  workspace,
  data_store,
  layer_name,
  title = NULL
)

Arguments

geoserver_url

A string, the base URL of the GeoServer instance.

user

A string, GeoServer username.

password

A string, GeoServer password.

workspace

A string, the name of the GeoServer workspace.

data_store

A string, the name of the data store where the layer resides.

layer_name

A string, the name of the layer to publish.

title

A string, an optional title for the layer. Defaults to the layer name if not provided.

Value

An integer indicating the status of the operation: - `0`: The layer was published successfully. - `1`: An error occurred during the publishing process.

See also

Examples

if (FALSE) { # \dontrun{
gs_publish_layer(
  geoserver_url = "http://my-geoserver.com/geoserver",
  user = "admin",
  password = "mypassword",
  workspace = "my_workspace",
  data_store = "my_datastore",
  layer_name = "my_layer",
  title = "My Layer Title"
)
} # }