Write Raster Bands to PostGIS

Writes each band of a raster (`terra::SpatRaster`) to a specified schema in a PostGIS database. Each band is written as a separate table in the database.

Usage

pg_write_bands(sr, conn, schema = "public", prefix = NULL, postfix = NULL)

Arguments

sr

A `terra::SpatRaster` object containing the raster bands to write.

conn

A database connection object to a PostGIS database (e.g., from `RPostgres::dbConnect`).

schema

A string specifying the schema in the PostGIS database where the raster layers will be stored. Default is `"public"`.

prefix

A string to prepend to each layer name. Default is `NULL`.

postfix

A string to append to each layer name. Default is `NULL`.

Value

The input `terra::SpatRaster` object, returned invisibly.

Details

Transforms the table name according to the Snake Case convention.

Examples

if (FALSE) { # \dontrun{
conn <- DBI::dbConnect(
  RPostgres::Postgres(),
  dbname = "mydb",
  host = "localhost",
  user = "user",
  password = "password"
)

sr <- terra::rast(nrows = 10, ncols = 10, nlyrs = 3, vals = runif(300))
pg_write_bands(sr, conn, schema = "geodata", prefix = "example_", postfix = "_raster")
DBI::dbDisconnect(conn)
} # }