Write GeoPackage Layers with Geometry to PostGIS

Transfers all vector layers with valid geometries from a GeoPackage file to a specified PostGIS database schema. Optionally allows setting a custom geometry column name, adding prefixes or postfixes to the table names, and renaming the layer fields to follow the Snake Case convention.

Usage

pg_write_layers(
  gpkg,
  conn,
  schema = "public",
  prefix = NULL,
  postfix = NULL,
  geom_colum = "geom",
  snake_case_fields = TRUE
)

Arguments

gpkg

A string, the path to the GeoPackage file.

conn

A PostGIS database connection object created with [RPostgres::dbConnect()].

schema

A string, the schema in PostGIS where layers will be stored. Default is `"public"`.

prefix

A string, an optional prefix to add to the table names in PostGIS. Default is `NULL`.

postfix

A string, an optional postfix to add to the table names in PostGIS. Default is `NULL`.

geom_colum

A string, the name of the geometry column to set. Default is `"geom"`.

snake_case_fields

A logical, whether to convert field names to Snake Case. Default is `TRUE`.

Value

Invisibly returns a character vector of the names of the tables written to PostGIS.

See also

Examples

if (FALSE) { # \dontrun{
conn <- DBI::dbConnect(RPostgres::Postgres(), dbname = "mydb")
gpkg_file <- "example.gpkg"
pg_write_layers(
  gpkg_file, conn, schema = "my_schema", prefix = "pre_", postfix = "_post"
)
DBI::dbDisconnect(conn)
} # }